[DEV-8] Auth: split to Android and Darwin
