+++ Sources/HumanRatedAI/Skip/skip.yml
@@ -11,10 +11,12 @@ build:
         - Token
     - block: 'dependencies'
       contents:
+        - 'implementation("com.auth0.android:jwtdecode:2.0.1")'
         - 'implementation("com.google.android.gms:play-services-auth:20.7.0")'
         - 'implementation("com.google.firebase:firebase-auth:22.3.0")'
         - 'implementation("com.google.firebase:firebase-core:21.1.1")'
-
+        - 'implementation("com.squareup.okhttp3:okhttp:4.10.0")'
+        
 # Add resource placeholders
 resources:
   contents:
